refactor: extract PassthroughInputFeature and PassthroughPreprocModule to their own module
Previously defined inline at the bottom of base_feature.py behind a factory
function, making them invisible to contributors following the normal feature
pattern. Moving them to passthrough_feature.py gives the class a discoverable
home alongside the other feature modules.

The factory create_passthrough_input_feature() moves to the same module; the
import in models/base.py is updated to the new path.
